requirements.txt
selenium
python-dotenv
pandas
numpy<=1.24.3
bs4
webdriver_manager
scikit-learn
typing-extensions
playwright
gitpython
docx2pdf
Django>=4.0
PyPDF2>=3.0
psycopg2>=2.9
pillow 
pytesseract
flask
flask-cors
werkzeug
wheel
image


